beginnings of the unit tests
