+++ ui/src/pages/DownloadPage/useDownloadPage.ts
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { UpdateDocumentResponse } from '../../utils/api.ts';
-import { generateCsvData } from './downloadPageController.ts';
+import { downloadCSV, downloadJSON } from './downloadPageController.ts';
 
 export function useDownloadPage() {
   const [verifiedData] = useState<UpdateDocumentResponse>(() => {
@@ -8,7 +8,7 @@ export function useDownloadPage() {
     return storedData ? JSON.parse(storedData)?.updated_document : {};
   });
 
-  function handleDownloadSubmit(event: React.FormEvent<HTMLFormElement>) {
+  async function handleDownloadSubmit(event: React.FormEvent<HTMLFormElement>) {
     event.preventDefault();
 
     if (!verifiedData.extracted_data) {
@@ -35,9 +35,9 @@ export function useDownloadPage() {
 
     // download function based on the selected file type
     if (fileType === 'csv') {
-      downloadCSV(verifiedData);
+      await downloadCSV(verifiedData);
     } else {
-      downloadJSON(verifiedData);
+      await downloadJSON(verifiedData);
     }
   }
 
@@ -45,36 +45,4 @@ export function useDownloadPage() {
     handleDownloadSubmit,
     verifiedData,
   };
-}
-
-function downloadCSV(verifiedData: UpdateDocumentResponse) {
-  const csvContent = generateCsvData(verifiedData.extracted_data ?? {});
-
-  downloadData(csvContent, 'text/csv', 'document.csv');
-}
-
-function downloadJSON(verifiedData: UpdateDocumentResponse) {
-  const jsonContent = JSON.stringify(verifiedData, null, 2);
-
-  downloadData(jsonContent, 'application/json', 'document.json');
-}
-
-async function downloadData(
-  content: string,
-  contentType: string,
-  filename: string
-) {
-  const blob = new Blob([content], { type: contentType });
-  const url = URL.createObjectURL(blob);
-
-  try {
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = filename;
-    link.click();
-  } finally {
-    setTimeout(() => {
-      URL.revokeObjectURL(url);
-    }, 1000); // Small delay to ensure download starts
-  }
 }
